src/app/components/admin/master/qualifications/route/qualifications-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IQualifications, Qualifications } from '../qualifications.model';
import { QualificationsService } from '../service/qualifications.service';

import { QualificationsRoutingResolveService } from './qualifications-routing-resolve.service';

describe('Qualifications routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: QualificationsRoutingResolveService;
  let service: QualificationsService;
  let resultQualifications: IQualifications | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(QualificationsRoutingResolveService);
    service = TestBed.inject(QualificationsService);
    resultQualifications = undefined;
  });

  describe('resolve', () => {
    it('should return IQualifications returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultQualifications = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultQualifications).toEqual({ id: 'ABC' });
    });

    it('should return new IQualifications if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultQualifications = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultQualifications).toEqual(new Qualifications());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Qualifications })));
      mockActivatedRouteSnapshot.params = { id: 'ABC' };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultQualifications = result;
      });

      // THEN
      expect(service.find).toBeCalledWith('ABC');
      expect(resultQualifications).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
